added libpng

+++ mat.h
@@ -2,13 +2,13 @@
 #define MATH
 
 #include <math.h>
+#include <stdint.h>
 #include <stdbool.h>
 
-
 typedef struct _mat_vec3 {
-    float x;
-    float y;
-    float z;
+    int64_t x;
+    int64_t y;
+    int64_t z;
 } mat_vec3;
 
 
@@ -17,7 +17,7 @@ float mat_vec3_abs(const mat_vec3* a) {
     return fabs(a->x * a->x + a->y * a->y + a->z * a->z);
 }
 
-mat_vec3 mat_vec3_smul(const mat_vec3* a, float k) {
+mat_vec3 mat_vec3_smul(const mat_vec3* a, int64_t k) {
     mat_vec3 res = {
         .x = k * a->x,
         .y = k * a->y,
@@ -27,7 +27,7 @@ mat_vec3 mat_vec3_smul(const mat_vec3* a, float k) {
 }
 
 mat_vec3 mat_vec3_inv(const mat_vec3 *a) {
-    return mat_vec3_smul(a, -1.0);
+    return mat_vec3_smul(a, -1);
 }
 
 // two vectors
